add jsdocs + use URLSearchParams + rm old comments
This commit adds documentation in the style of jsdocs. Nice! To do this,
I added `console.log` calls to print the variables that were going into
and out of the function. That allowed me to figure out the types for
arguments and the structure of the returned object. To describe the
returned object, I used a typedef. This typedef defines one item in the
returns object, and the returned object includes an array of these
items.

I have several questions after working on this function.
1. How should we handle calls to `fetch`? In this function, we check
   whether the output is an empty string. But would it ever be? Would it
   make more sense to check the response's return code? Or perhaps use a
   try-catch and log errors?
2. We return an array if `fetch` returns an empty string. Do we do
   anything with this array? It seems like a `console.log` would suffice
   here. Actually, I will look into the call stack and probably change
   this to a `console.log`.