set url parameters with selected queries
When a user selects an option, the option will be set as a parameter in
the url. The next step is to set the initial values for the select2
boxes using the url parameters, if they are given. Once this is
implemented, a user can copy-paste a url to select the same things.
Imagine that a user wants to send the results to a friend. That friend
can enter the url and will have the same selection as the original user.